pyproject.toml
[project]
name = "boot-agent-speedrun"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.21.1",
    "python-dotenv==1.1.0",
]
